Move PulseDocument to a separate file
