+++ Sources/PulseCore/Helpers/CoreData+Extensions.swift
@@ -100,3 +100,46 @@ extension NSPersistentStoreCoordinator {
         )
     }
 }
+
+extension NSEntityDescription {
+    convenience init<T>(name: String, class: T.Type) where T: NSManagedObject {
+        self.init()
+        self.name = name
+        self.managedObjectClassName = T.self.description()
+    }
+}
+
+extension NSAttributeDescription {
+    convenience init(name: String, type: NSAttributeType, _ configure: (NSAttributeDescription) -> Void = { _ in }) {
+        self.init()
+        self.name = name
+        self.attributeType = type
+        configure(self)
+    }
+}
+
+enum RelationshipType {
+    case oneToMany
+    case oneToOne(isOptional: Bool = false)
+}
+
+extension NSRelationshipDescription {
+    static func make(name: String,
+                     type: RelationshipType,
+                     deleteRule: NSDeleteRule = .cascadeDeleteRule,
+                     entity: NSEntityDescription) -> NSRelationshipDescription {
+        let relationship = NSRelationshipDescription()
+        relationship.name = name
+        relationship.deleteRule = deleteRule
+        relationship.destinationEntity = entity
+        switch type {
+        case .oneToMany:
+            relationship.maxCount = 0
+            relationship.minCount = 0
+        case .oneToOne(let isOptional):
+            relationship.maxCount = 1
+            relationship.minCount = isOptional ? 0 : 1
+        }
+        return relationship
+    }
+}
