+++ Sources/PulseCore/LoggerStore/LoggerStore+Model.swift
@@ -82,134 +82,4 @@ extension LoggerStore {
         model.entities = [message, metadata, request, requestProgress, blob, inlinedData]
         return model
     }()
-
-    static let archiveModel: NSManagedObjectModel = {
-        let model = NSManagedObjectModel()
-
-        let archive = NSEntityDescription(name: "PulseDocumentEntity", class: PulseDocumentEntity.self)
-        let blob = NSEntityDescription(name: "PulseBlobEntity", class: PulseBlobEntity.self)
-
-        archive.properties = [
-            NSAttributeDescription(name: "info", type: .binaryDataAttributeType),
-            NSAttributeDescription(name: "database", type: .binaryDataAttributeType),
-            NSRelationshipDescription.make(name: "blobs", type: .oneToMany, entity: blob)
-        ]
-
-        blob.properties = [
-            NSAttributeDescription(name: "key", type: .stringAttributeType),
-            NSAttributeDescription(name: "data", type: .binaryDataAttributeType)
-        ]
-
-        model.entities = [archive, blob]
-
-        return model
-    }()
-}
-
-final class PulseDocumentEntity: NSManagedObject {
-    @NSManaged var info: Data
-    @NSManaged var database: Data
-    @NSManaged var blobs: Set<PulseBlobEntity>
-}
-
-final class PulseBlobEntity: NSManagedObject {
-    @NSManaged var key: String
-    @NSManaged var data: Data
-}
-
-final class PulseDocument {
-    let container: NSPersistentContainer
-    var context: NSManagedObjectContext { container.viewContext }
-
-    private func document() throws -> PulseDocumentEntity {
-        guard let document = self._document else {
-            throw LoggerStore.Error.unknownError // Programmatic error
-        }
-        return document
-    }
-
-    private var _document: PulseDocumentEntity?
-
-    init(documentURL: URL) throws {
-        #warning("move archive model jer")
-        self.container = NSPersistentContainer(name: documentURL.lastPathComponent, managedObjectModel: LoggerStore.archiveModel)
-        let store = NSPersistentStoreDescription(url: documentURL)
-        store.setValue("NONE" as NSString, forPragmaNamed: "journal_mode")
-        store.setOption(true as NSNumber, forKey: NSSQLiteManualVacuumOption)
-        container.persistentStoreDescriptions = [store]
-
-        try container.loadStore()
-    }
-
-    /// Opens existing store and returns store info.
-    func open() throws -> LoggerStore.Info {
-        guard let document = try context.first(PulseDocumentEntity.self) else {
-            throw LoggerStore.Error.storeInvalid
-        }
-        self._document = document
-        return try JSONDecoder().decode(LoggerStore.Info.self, from: document.info)
-    }
-
-    // Opens an existing database.
-    func database() throws -> Data {
-        try document().database.decompressed()
-    }
-
-    func getBlob(forKey key: String) -> Data? {
-        try? context.first(LoggerBlobHandleEntity.self) {
-            $0.predicate = NSPredicate(format: "key == %@", key)
-        }?.data
-    }
-
-    func close() throws {
-        let coordinator = container.persistentStoreCoordinator
-        for store in coordinator.persistentStores {
-            try coordinator.remove(store)
-        }
-    }
-}
-
-// MARK: - Helpers
-
-private extension NSEntityDescription {
-    convenience init<T>(name: String, class: T.Type) where T: NSManagedObject {
-        self.init()
-        self.name = name
-        self.managedObjectClassName = T.self.description()
-    }
-}
-
-private extension NSAttributeDescription {
-    convenience init(name: String, type: NSAttributeType, _ configure: (NSAttributeDescription) -> Void = { _ in }) {
-        self.init()
-        self.name = name
-        self.attributeType = type
-        configure(self)
-    }
-}
-
-private enum RelationshipType {
-    case oneToMany
-    case oneToOne(isOptional: Bool = false)
-}
-
-private extension NSRelationshipDescription {
-    static func make(name: String,
-                     type: RelationshipType,
-                     deleteRule: NSDeleteRule = .cascadeDeleteRule,
-                     entity: NSEntityDescription) -> NSRelationshipDescription {
-        let relationship = NSRelationshipDescription()
-        relationship.name = name
-        relationship.deleteRule = deleteRule
-        relationship.destinationEntity = entity
-        switch type {
-        case .oneToMany:
-            relationship.maxCount = 0
-            relationship.minCount = 0
-        case .oneToOne(let isOptional):
-            relationship.maxCount = 1
-            relationship.minCount = isOptional ? 0 : 1
-        }
-        return relationship
-    }
 }
